refactor(cloud): separate JSX into components
resolves #28
